Move FieldData and DictFieldData (nee DictModel) into a separate module
