+++ xblock/fields.py
@@ -7,90 +7,7 @@ storage mechanism is.
 
 import copy
 from collections import namedtuple
-
-UNSET = object()
-
-
-class FieldData(object):
-    """
-    An interface allowing access to an XBlock's field values indexed by field names
-    """
-    def get(self, block, name, default=UNSET):
-        """
-        Retrieve the value for the field named `name` for the XBlock `block`.
-
-        If a value is provided for `default`, then it will be
-        returned if no value is set
-
-        The value returned may be mutated without modifying the backing store
-
-        :param block: block to inspect
-        :type block: :class:`~xblock.core.XBlock`
-        :param name: field name to look up
-        :type name: str
-        """
-        raise NotImplementedError
-
-    def set(self, block, name, value):
-        """
-        Set the value of the field named `name` for XBlock `block`.
-
-        `value` may be mutated after this call without affecting the backing store
-
-        :param block: block to modify
-        :type block: :class:`~xblock.core.XBlock`
-        :param name: field name to set
-        :type name: str
-        :param value: value to set
-        """
-        raise NotImplementedError
-
-    def delete(self, block, name):
-        """
-        Reset the value of the field named `name` to the default for XBlock `block`.
-
-        :param block: block to modify
-        :type block: :class:`~xblock.core.XBlock`
-        :param name: field name to delete
-        :type name: str
-        """
-        raise NotImplementedError
-
-    def has(self, block, name):
-        """
-        Return whether or not the field named `name` has a non-default value for the XBlock `block`
-
-        :param block: block to check
-        :type block: :class:`~xblock.core.XBlock`
-        :param name: field name
-        :type name: str
-        """
-        raise NotImplementedError
-
-    def set_many(self, block, update_dict):
-        """
-        Update many fields on an XBlock simultaneously.
-
-        :param block: the block to update
-        :type block: :class:`~xblock.core.XBlock`
-        :param update_dict: A map of field names to their new values
-        :type update_dict: dict
-        """
-        for key, value in update_dict.items():
-            self.set(block, key, value)
-
-    def default(self, block, name):
-        """
-        Get the default value for this field which may depend on context or may just be the field's global
-        default. The default behavior is to raise KeyError which will cause the caller to return the field's
-        global default.
-
-        :param block: the block containing the field being defaulted
-        :type block: :class:`~xblock.core.XBlock`
-        :param name: the field's name
-        :type name: `str`
-        """
-        raise KeyError(name)
+from xblock.field_data import UNSET
 
 
 class BlockScope(object):
